v2: Breaking changes from v1 to v2.0.0 (#415)
