+++ algosdk/__init__.py
@@ -1,16 +1,13 @@
 from . import abi
 from . import account
-from . import algod
 from . import auction
 from . import constants
 from . import dryrun_results
 from . import encoding
 from . import error
-from . import future
 from . import kmd
 from . import logic
 from . import mnemonic
-from . import template
 from . import transaction
 from . import util
 from . import v2client
