+++ algosdk/atomic_transaction_composer.py
@@ -13,9 +13,8 @@ from typing import (
     cast,
 )
 
-from algosdk import abi, error
+from algosdk import abi, error, transaction
 from algosdk.abi.address_type import AddressType
-from algosdk.future import transaction
 from algosdk.v2client import algod
 
 # The first four bytes of an ABI method call return must have this hash
