+++ algosdk/encoding.py
@@ -1,12 +1,11 @@
 import base64
-import warnings
 from collections import OrderedDict
 from typing import Union
 
 import msgpack
 from Cryptodome.Hash import SHA512
 
-from algosdk import auction, constants, error, future, transaction
+from algosdk import auction, constants, error, transaction
 
 
 def msgpack_encode(obj):
@@ -56,51 +55,8 @@ def _sort_dict(d):
     return od
 
 
-def future_msgpack_decode(enc):
-    """
-    Decode a msgpack encoded object from a string.
-
-    Args:
-        enc (str): string to be decoded
-
-    Returns:
-        Transaction, SignedTransaction, Multisig, Bid, or SignedBid:\
-            decoded object
-    """
-    decoded = enc
-    if not isinstance(enc, dict):
-        decoded = msgpack.unpackb(base64.b64decode(enc), raw=False)
-    if "type" in decoded:
-        return future.transaction.Transaction.undictify(decoded)
-    if "l" in decoded:
-        return future.transaction.LogicSig.undictify(decoded)
-    if "msig" in decoded:
-        return future.transaction.MultisigTransaction.undictify(decoded)
-    if "lsig" in decoded:
-        if "txn" in decoded:
-            return future.transaction.LogicSigTransaction.undictify(decoded)
-        return future.transaction.LogicSigAccount.undictify(decoded)
-    if "sig" in decoded:
-        return future.transaction.SignedTransaction.undictify(decoded)
-    if "txn" in decoded:
-        return future.transaction.Transaction.undictify(decoded["txn"])
-    if "subsig" in decoded:
-        return future.transaction.Multisig.undictify(decoded)
-    if "txlist" in decoded:
-        return future.transaction.TxGroup.undictify(decoded)
-    if "t" in decoded:
-        return auction.NoteField.undictify(decoded)
-    if "bid" in decoded:
-        return auction.SignedBid.undictify(decoded)
-    if "auc" in decoded:
-        return auction.Bid.undictify(decoded)
-
-
 def msgpack_decode(enc):
     """
-    NOTE: This method is deprecated:
-    Please use `future_msgpack_decode` instead.
-
     Decode a msgpack encoded object from a string.
 
     Args:
@@ -110,11 +66,6 @@ def msgpack_decode(enc):
         Transaction, SignedTransaction, Multisig, Bid, or SignedBid:\
             decoded object
     """
-    warnings.warn(
-        "`msgpack_decode` is being deprecated. "
-        "Please use `future_msgpack_decode` instead.",
-        DeprecationWarning,
-    )
     decoded = enc
     if not isinstance(enc, dict):
         decoded = msgpack.unpackb(base64.b64decode(enc), raw=False)
@@ -125,7 +76,9 @@ def msgpack_decode(enc):
     if "msig" in decoded:
         return transaction.MultisigTransaction.undictify(decoded)
     if "lsig" in decoded:
-        return transaction.LogicSigTransaction.undictify(decoded)
+        if "txn" in decoded:
+            return transaction.LogicSigTransaction.undictify(decoded)
+        return transaction.LogicSigAccount.undictify(decoded)
     if "sig" in decoded:
         return transaction.SignedTransaction.undictify(decoded)
     if "txn" in decoded:
