--- algosdk/future/__init__.py
@@ -1,4 +1,0 @@
-from . import template
-from . import transaction
-
-name = "future"
