+++ algosdk/kmd.py
@@ -4,7 +4,7 @@ import urllib.error
 from urllib import parse
 from urllib.request import Request, urlopen
 
-from . import constants, encoding, error, future
+from . import constants, encoding, error, transaction
 
 api_version_path_prefix = "/v1"
 
@@ -323,7 +323,7 @@ class KMDClient:
             query["public_key"] = signing_address
         result = self.kmd_request("POST", req, data=query)
         result = result["signed_transaction"]
-        return encoding.future_msgpack_decode(result)
+        return encoding.msgpack_decode(result)
 
     def list_multisig(self, handle):
         """
@@ -381,7 +381,7 @@ class KMDClient:
         result = self.kmd_request("POST", req, data=query)
         pks = result["pks"]
         pks = [encoding.encode_address(base64.b64decode(p)) for p in pks]
-        msig = future.transaction.Multisig(
+        msig = transaction.Multisig(
             result["multisig_version"], result["threshold"], pks
         )
         return msig
@@ -434,6 +434,6 @@ class KMDClient:
             "partial_multisig": partial,
         }
         result = self.kmd_request("POST", req, data=query)["multisig"]
-        msig = encoding.future_msgpack_decode(result)
+        msig = encoding.msgpack_decode(result)
         mtx.multisig = msig
         return mtx
