--- docs/algosdk/algod.rst
@@ -1,7 +1,0 @@
-algod
-=====
-
-.. automodule:: algosdk.algod
-   :members:
-   :undoc-members:
-   :show-inheritance:
