--- docs/algosdk/future/index.rst
@@ -1,8 +1,0 @@
-future
-======
-.. toctree::
-   :maxdepth: 10
-
-   template
-   transaction
-
