--- docs/algosdk/future/template.rst
@@ -1,7 +1,0 @@
-future.template
-===============
-
-.. automodule:: algosdk.future.template
-   :members:
-   :undoc-members:
-   :show-inheritance:
