--- docs/algosdk/future/transaction.rst
@@ -1,7 +1,0 @@
-future.transaction
-==================
-
-.. automodule:: algosdk.future.transaction
-   :members:
-   :undoc-members:
-   :show-inheritance:
