--- docs/algosdk/template.rst
@@ -1,7 +1,0 @@
-template
-========
-
-.. automodule:: algosdk.template
-   :members:
-   :undoc-members:
-   :show-inheritance:
