+++ examples/log_sig_example.py
@@ -1,9 +1,8 @@
 # Example: creating a LogicSig transaction signed by a program that never approves the transfer.
 
 import tokens
-from algosdk import account
+from algosdk import account, transaction
 from algosdk.v2client import algod
-from algosdk.future import transaction
 
 program = b"\x01\x20\x01\x00\x22"  # int 0
 lsig = transaction.LogicSigAccount(program)
